requirements.txt
flask
pandas
torch
Pillow
open-clip-torch
scikit-learn
numpy